Post-merge fixes
